Created Repository pattern
There is currently no access to any database yet, but the pattern is
designed.
The Repository abstract class will likely take a <tableName: string> as
a parameter in its constructor, but it's removed for now.
Each feature-specific Repository will have its own feature-specific
methods, while extending the Repository abstract class.
The Post model was established, but it is only temporary for testing.
The current Post model is only partially complete.

More on this tomorrow.